requirements/devel.txt
coverage==3.5.1
django-coverage==1.2.1
pysqlite==2.6.3
